+++ crates/teloxide-core/src/payloads/copy_message.rs
@@ -7,9 +7,11 @@ use crate::types::{
 };
 
 impl_payload! {
-    /// Use this method to copy messages of any kind. The method is analogous to the method forwardMessage, but the copied message doesn't have a link to the original message. Returns the [`MessageId`] of the sent message on success.
+    /// Use this method to copy messages of any kind. Service messages, paid media messages, giveaway messages, giveaway winners messages, and invoice messages can't be copied. A quiz [`Poll`] can be copied only if the value of the field *correct_option_id* is known to the bot. The method is analogous to the method [`ForwardMessage`], but the copied message doesn't have a link to the original message. Returns the [`MessageId`] of the sent message on success.
     ///
+    /// [`Poll`]: crate::types::Poll
     /// [`MessageId`]: crate::types::MessageId
+    /// [`ForwardMessage`]: crate::payloads::ForwardMessage
     #[derive(Debug, PartialEq, Eq, Hash, Clone, Serialize)]
     pub CopyMessage (CopyMessageSetters) => MessageId {
         required {