+++ crates/teloxide-core/src/requests/multipart_payload.rs
@@ -1,7 +1,7 @@
 use crate::{
     payloads,
     requests::Payload,
-    types::{InputFile, InputFileLike, InputMedia, InputSticker},
+    types::{InputFile, InputFileLike, InputMedia, InputPaidMedia, InputSticker},
 };
 
 /// Payloads that need to be sent as `multipart/form-data` because they contain
@@ -10,6 +10,16 @@ pub trait MultipartPayload: Payload {
     fn copy_files(&self, into: &mut dyn FnMut(InputFile));
 
     fn move_files(&mut self, into: &mut dyn FnMut(InputFile));
+}
+
+impl MultipartPayload for payloads::SendPaidMedia {
+    fn copy_files(&self, into: &mut dyn FnMut(InputFile)) {
+        self.media.iter().flat_map(InputPaidMedia::files).for_each(|f| f.copy_into(into))
+    }
+
+    fn move_files(&mut self, into: &mut dyn FnMut(InputFile)) {
+        self.media.iter_mut().flat_map(InputPaidMedia::files_mut).for_each(|f| f.move_into(into))
+    }
 }
 
 impl MultipartPayload for payloads::SendMediaGroup {