+++ crates/teloxide-core/src/types/external_reply_info.rs
@@ -2,8 +2,8 @@ use serde::{Deserialize, Serialize};
 
 use crate::types::{
     Animation, Audio, Chat, Contact, Dice, Document, Game, Giveaway, GiveawayWinners, Invoice,
-    LinkPreviewOptions, Location, MessageId, MessageOrigin, PhotoSize, Poll, Sticker, Story, Venue,
-    Video, VideoNote, Voice,
+    LinkPreviewOptions, Location, MessageId, MessageOrigin, PaidMediaInfo, PhotoSize, Poll,
+    Sticker, Story, Venue, Video, VideoNote, Voice,
 };
 
 /// This object contains information about a message that is being replied to,
@@ -46,6 +46,7 @@ pub enum ExternalReplyInfoKind {
     Contact(Contact),
     Dice(Dice),
     Document(Document),
+    PaidMedia(PaidMediaInfo),
     Game(Game),
     Venue(Venue),
     Location(Location),